remove the non-capability-enabled notificator and make the new one more capability-aware
