+++ src/dataflow/operators/unary.rs
@@ -7,7 +7,7 @@ use std::default::Default;
 use progress::nested::subgraph::{Source, Target};
 
 use progress::count_map::CountMap;
-use dataflow::operators::CapabilityNotificator as Notificator;
+use dataflow::operators::Notificator;
 use progress::{Timestamp, Operate, Antichain};
 use dataflow::channels::pushers::Tee;
 use dataflow::channels::pushers::Counter as PushCounter;
@@ -153,16 +153,14 @@ Operator<T, D1, D2, L> {
                notify:   Option<(Vec<T>, usize)>)
            -> Operator<T, D1, D2, L> {
 
-        let internal = Rc::new(RefCell::new(CountMap::new()));
-
         Operator {
             name:    name,
             input:       receiver,
             output:      PushBuffer::new(PushCounter::new(targets, Rc::new(RefCell::new(CountMap::new())))),
-            notificator: Notificator::new(internal.clone()),
+            notificator: Notificator::new(),
             logic:   logic,
             notify:  notify,
-            internal_changes: internal.clone(),
+            internal_changes: Rc::new(RefCell::new(CountMap::new())),
         }
     }
 }
